requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
Babel==2.9.1
backcall==0.2.0
baidu-aip==2.2.18.0
bce-python-sdk==0.8.61
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
cachetools==4.2.2
cchardet==2.1.7
certifi==2020.6.20
cffi==1.14.3
cfgv==3.3.0
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
comtypes==1.1.10
configobj==5.0.6
configparser==5.0.1
croniter==1.0.15
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
dill==0.3.4
distlib==0.3.1
docx==0.2.4
docxcompose==1.3.1
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
etelemetry==0.2.2
fake-useragent==0.1.11
Faker==8.10.3
filelock==3.0.12
fire==0.4.0
fitz==0.0.1.dev2
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
fonttools==4.24.4
frontend==0.0.3
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.20
gopup==0.2.8
h11==0.12.0
h5py==3.3.0
html2text==2020.1.16
httplib2==0.19.0
identify==2.2.11
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
img2pdf==0.4.1
imgaug==0.4.0
iniconfig==1.1.1
install==1.3.4
iopath==0.1.9
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.6.4
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.17.2
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-kite==1.1.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.2.0
layoutparser @ https://paddleocr.bj.bcebos.com/whl/layoutparser-0.0.0-py3-none-any.whl
lazy-object-proxy==1.4.3
levenshtein==0.12.0
lmdb==1.2.1
lsp==0.1.1
lxml==4.5.2
MarkupSafe==2.0.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
moviepy==1.0.3
multidict==5.1.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
nodeenv==1.6.0
notebook==6.1.4
numpy==1.19.2
OpenCC==1.1.1
opencv-contrib-python==4.4.0.46
opencv-python==4.5.2.54
openpyxl==3.0.5
packaging==20.4
paddle2onnx==0.7
paddlehub==2.1.0
paddlenlp==2.0.7
paddleocr==2.2.0.2
paddlepaddle==2.1.2
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathlib==1.0.1
pdf2docx==0.5.2
pdfkit==0.6.1
pdfminer==20191125
pdfminer.six==20200517
pdfminer3k==1.3.4
pdfplumber==0.5.28
pickleshare==0.7.5
pikepdf==2.16.1
Pillow==8.0.0
pipenv==2020.11.15
pluggy==0.13.1
ply==3.11
portalocker==2.3.0
pre-commit==2.13.0
premailer==3.10.0
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.17.3
prov==2.0.0
py==1.9.0
pyclipper==1.3.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.9
pydbgen==1.0.5
pydot==1.4.1
PyExecJS==1.5.1
pyflakes==2.3.1
pyforest==1.0.3
Pygments==2.7.1
pylint==2.6.0
PyMuPDF==1.18.14
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
pyquery==1.4.3
pyrsistent==0.17.3
PySide2==5.15.2
PySocks==1.7.1
pytesseract==0.3.8
pytest==6.1.2
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-docx==0.8.10
pyttsx3==2.90
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyxnat==1.4
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
rarfile==4.0
rdflib==5.0.0
regex==2021.4.4
reload==0.9
removebg==0.4
requests==2.24.0
retrying==1.3.3
rich==10.7.0
scihub==0.0.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
seqeval==1.2.2
Shapely==1.7.1
shellcheck-py==0.7.2.1
shiboken2==5.15.2
simplejson==3.17.2
six==1.15.0
smmap==4.0.0
sortedcontainers==2.4.0
soupsieve==2.0.1
sqlparse==0.4.1
starlette==0.14.2
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.1
torch==1.9.0
torchvision==0.10.0
tornado==6.0.4
tqdm==4.59.0
traitlets==5.0.5
traits==6.2.0
turtle==0.0.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.22
useragent==0.1.1
uvicorn==0.13.3
virtualenv==20.4.6
virtualenv-clone==0.5.4
visualdl==2.2.0
w3lib==1.22.0
Wand==0.6.6
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
wechatsogou==4.5.4
Werkzeug==2.0.1
widgetsnbextension==3.5.1
win10toast==0.9
wkhtmltopdf==0.2
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlutils==2.0.0
xlwt==1.3.0
yapf==0.31.0
yarl==1.6.3


